A bit of refactoring domain_ownership_service to abstraction
