+++ app/lib/services/vault_service.rb
@@ -1,5 +1,7 @@
 module Services
   class VaultService
+    attr_reader :client
+
     def initialize
       # TODO create a new token for use in the session
       @client = Vault::Client.new(
@@ -11,7 +13,7 @@ module Services
     def issue_cert(cert_issue_request)
       opts = cert_issue_request.attributes
       # Generate the TLS certificate using the intermediate CA
-      tls_cert = @client.logical.write(Rails.configuration.astral[:vault_cert_path], opts)
+      tls_cert = client.logical.write(Rails.configuration.astral[:vault_cert_path], opts)
       OpenStruct.new tls_cert.data
     end
   end
